Fix calculation logic and reuse existing listing
